Revert the change to the cast in dispatchAsyncMainThreadWithDelay. Navi can't remember what warning this was fixing, and now I'm not getting it.
Casting to unsigned long doesn't feel right because the dispatchAsyncMainThreadWithDelay
function is declared to take int.  In particular, we don't want weird sign-extension
problems if the given delay is negative.

+++ Tidbits/Dispatch.h
@@ -30,6 +30,6 @@ void dispatchAsyncBackgroundThreadWithDelay(int delay_msec, dispatch_queue_prior
 #if DURATION_WARNING_ENABLED
 #define dispatchSyncMainThread(...) dispatchSyncMainThread(__PRETTY_FUNCTION__, __LINE__, ## __VA_ARGS__)
 #define dispatchAsyncMainThread(...) dispatchAsyncMainThread(__PRETTY_FUNCTION__, __LINE__, ## __VA_ARGS__)
-#define dispatchAsyncMainThreadWithDelay(__delay, ...) dispatchAsyncMainThreadWithDelay(__PRETTY_FUNCTION__, __LINE__, (unsigned long)__delay, ## __VA_ARGS__)
+#define dispatchAsyncMainThreadWithDelay(__delay, ...) dispatchAsyncMainThreadWithDelay(__PRETTY_FUNCTION__, __LINE__, (int)__delay, ## __VA_ARGS__)
 #define dispatchSyncMainThreadWithResult(...) dispatchSyncMainThreadWithResult(__PRETTY_FUNCTION__, __LINE__, ## __VA_ARGS__)
 #endif
